// aiot-frontend/src/api/generated/device-service/models/product-v1-product.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Product message
 */
export interface ProductV1Product {
    'id'?: string;
    'productKey'?: string;
    'name'?: string;
    'description'?: string;
    'category'?: string;
    'status'?: string;
    'metadata'?: string;
    'deviceCount'?: number;
    'createdAt'?: string;
    'updatedAt'?: string;
    'nodeType'?: string;
    'connectivityMethod'?: string;
    'accessProtocol'?: string;
    'tenantId'?: string;
    'deletedAt'?: string;
}